requirements.txt
bcrypt==3.2.0
cffi==1.14.4
Django==2.2.7
django-filter==2.4.0
pycparser==2.20
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
